requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lark-oapi
MarkupSafe==2.1.3
pycryptodome==3.18.0
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
python-dotenv
bs4
Markdown
openai
pandas
langgraph
langchain-community
langchain-anthropic
tavily-python
langchain_openai